requirements.txt
raylib
Pillow